requirements/requirements.txt
flask==0.12
flask_sqlalchemy==2.3.0
psycopg2==2.7.3
pymongo==3.2.2
marshmallow>=3.0.0b3
requests==2.18
gunicorn==19.6.0
gevent==1.1.1
json-logging-py==0.2
